Substitute `first` implementation
To one based on `next()`.

On @lucyb's suggestion, so that we can demonstrate the implementations
are the same, based on the existing `first` package tests.

See https://github.com/opensafely-core/job-server/pull/4503#discussion_r1700521715

The `first` tests pass as of this commit.

We'll later inline the function, because this removes the indirection
and also allows us to simplify the usage where possible.