Add deployment administrator role
Job management through the UI is required for tech-support.
This can only be performed by developers with the deployment
administrator role, as defined in the developer permissions log.

For example, re-running user-submitted jobs after a system issue or
configuration change.

We believe the permissions defined here to be adequate for this role,
but will revisit if the experience of users assigned this role
reveals further requirements.