Remove user github access_tokens from db
PR #4539 prevented future storage of users's GitHub access tokens when
signing in, but left previously-stored tokens in the database.

Due to these tokens' short expiry time these residual tokens in the db
do not represent a security risk, but may confuse anyone looking at these
data in the future.

This migration removes the residual tokens so that the database state
is consistent with the current authentication pipeline.