Reduce use of `attrs`
This reduces the use of `attrs` replacing uses where a standard
`dataclass` is sufficient.

There are some slightly more complicated uses which are not changed,
but could be removed in future:

* the use of validation in `jobserver/runtime.py`; this may be replaceable
  with a `__post_init__` method or use of `pydantic`
* the use of `asdict()` to convert an `attrs` instance from
  `interactive-templates` in `interactive/submit.py`; this may be
  possible to update to use the `.asdict()` method on the instance,
  removing the direct dependency

Removing those would completely remove `attrs` as a direct dependency of
job-server. However, it would remain as a transitive dependency,
introduced by `interactive-templates`.