Enable migrations when running tests in CI
By default, we disable migrations when running the tests (see
`tool.pytest.ini_options.addopts` in `pyproject.toml`). The
pytest-django documentation suggests that doing so might improve
performance when there are "several" migrations; experience suggests
that doing so ensures that an error in a new migration doesn't cause
every test to fail (because every test has database access) and
(correctly) drops the database when the tests have run.

Clearly, the drawback is that unless we run `migrate` before deploying a
new migration, we will only discover an error when the new migration is
deployed. Ouch!

It's easy to forget to run `migrate`, but tedious to watch every test
fail locally. As a compromise, we enable migrations when running the
tests in CI. The alternative solutions are fiddly, as they involve
writing a test that drops, creates, and runs `migrate`; or overriding
pytest-django's `django_db_setup` fixture.