Change repo-project from one-one to one-many
For much, but not all, of job-server's life, one repo could be
associated with one project. However, there are now a small number of
repos that should be associated with two projects: these repos have
"legacy" and "transitional" projects. To accommodate these repos, we
change the repo-project relationship from one-to-one to one-to-many.

Are you expecting to see changes to a model? And not expecting to see
changes to a form? The repo-project relationship was actually
one-to-many early in job-server's life, so the database already contains
several repos that are associated with two projects. For this reason,
the one-to-one repo-project relationship was enforced within a form and
not within a model.

Are you concerned that we may no longer be supporting an important
use-case? The original use-case for changing from repo-project
relationship from one-to-many to one-one was (we think) to help a member
of the public navigate from GitHub's website to job-server's website
("When was this code run?"). In practice, however, supporting this
use-case isn't easy. We feel an alternative design is required.