Give some projects orgs
We would expect projects to have orgs in production, so we should expect
the same in testing. However, because we're nervous about what might
happen if we change `ProjectFactory`, and because we think the whole
org/orgs thing was a terrible mistake, we only give some projects orgs:
that is, only those projects used in tests that render some aspect of
those orgs.