Point to Airlock for viewing log outputs
There's no point linking directly to the relevant Airlock URL because we
don't expect users to be accessing Job Server from within Level 4 going
forward, and even you have Airlock open in a VM you can't copy/paste
URLs across.

Instead we link to the Airlock docs and describe the location of the
file. This replaces both the link the output viewer and the reference to
the old "Level 4 file sync" directory.

Previously we restricted these links to only show for the TPP backend,
but on the assumption that we will be running an instance of Airlock for
every backend I have removed this restriction.