Handle HTTP timeout
We already handle HTTP Errors from GitHub, this change makes the view
behave the same if and when we get a timeout error from the ThreadPoolExecutor.

We could attempt some retry logic here but as this is a non-critical
feature it's probably not worth it.

Instead, a placeholder Repo stating the unavailability of the GitHub API
is returned which should allow the template to render with this hopefully
useful message.