Add explicit dependencies on building to Docker commands
This seems redundant because `docker compose` will build the images as
necessary, but for that to succeed we need to pass in the necessary
build args as in the `docker:build` command. Adding the dependency
avoids having to duplicate the setting of those args or introducing
some fiddly abstractions to set them the same way in two places.