Separate out deploy and build/test workflows
This is done via a GitHub Actions reusable workflow using
`workflow_call` as the trigger.

This enables setting `concurrency` on a new deploy workflow, without
needing to copy-paste the long-ish build and test workflow there. The
tests are still run before deploy.

This should fix #4607, because the deploy workflow will only run once
during an in-progress deploy.

Prior to this change, it was possible for two runs of the same build,
test, deploy workflow to proceed at once, with only the deploy job being
limited to one simultaneous run.

That meant that if an earlier workflow run (1) and later workflow run
(2) were started close in time to each other, then run 1 might start
before run 2, but run 2 might actually reach the deploy job first. This
would mean run 2 would deploy before run 1, and the deployed state would
be associated with an earlier commit.

The only downside is that deploys in quick succession might take a bit
longer, because they can't run the tests in parallel.

Note that with GitHub's `concurrency`, it allows one active and one
queued job. A later queued job replaces the current queued job, so it's
possible that not every deploy happens, but I believe this was the case
with `concurrency` set on the old deploy job.