Use pure Django query to find redirects
We allow underscores in project and workspace names.

When finding redirects, we search for redirects whose `old_url` is at
the beginning of the requested URL.

Previous approaches hand-constructed the SQL WHERE clause with a LIKE
operator which was buggy due to underscores being LIKE wildcards.

With this approach, Django generates the WHERE clause with all the
necessary escaping of characters in `old_url` that could be interpreted
as wildcards.

Co-authored-by: Mike Kelly <mike.kelly@thedatalab.org>