Rename CoreDeveloper to StaffAreaAdministrator
This role controls access to the Staff Area,
role holders might not necessarily be  developers, but admin staff.

Therefore, Staff Area Administrator was deemed to be a better fit for this role

In order to support a data migration from CoreDeveloper to StaffAreaAdministrator,
a vestigial CoreDeveloper role remains.
This is to be removed once all dependent migrations are removed.