Extra instructions for changing the User model
We have a custom view that's used by Grafana to access a subset of the User
table. If fields on the User model are changed, this updates the underlying
database table linked to that view, and means the view needs to be regenerated.
I've added a pointer to the instructions that tells someone how to do that.