Assert `name` field is cleaned
This adds an assertion to `test_workspacecreateform_success` that
matches the assertion in
`test_workspacecreateform_success_with_mixed_case_name`. In both cases,
the assertions check that the `name` field is cleaned.

It wasn't clear to me why the assertion message was the cleaned `name`
field (possibly for debugging?), so I removed the assertion message.