Don't render large YAML in JobRequestDetail view.
The job-request-detail template doesn't display project YAML if the context flags
project_yaml is_too_large (from #3870). Therefore the view code doesn't need to
render the definition in that case. Doing so was slowing down page rendering by
5 or more seconds in some cases.

The extant code arbitrarily defined "large" as longer than 10,000 characters,
but the comments mentioned ~1MB , which doesn't match. I've updated the
comment. I've not considered updating the arbitrary value used.

As well as a unit test for the "too large" case, I added one for the rendering
actually taking effect, as it was missing. And added checks that the project_yaml
is_too_large or is_empty are as they should be in the other "happy path" tests.