Re-name --core-developer management command flag.
In #4604/#4639 the CoreDeveloper role was re-named to StaffAreaAdministrator.
Direct references were updated. The management command create_user still has a
vestigial indirect reference in the flag name --core-developer. Let's tidy it
up as the vestigial reference is potentially confusing and unclear. Better to
reference the new role name.

Add short name to --staff-area-administrator and --output-checker for
convenience when using the command interactively without autocomplete.