Use `form.data`, not `kwargs`. Validate more in tests.
POSTed data may be provided to a `Form` as either the first positional argument
or as the `data` key-word argument of its `__init__` function. Therefore we
should access it through self.data, not through `kwargs`. Otherwise `data` will
be ignored in our code if passed by position, which doesn't follow the `Form`
API and is surprising.

This hasn't been noticed so far because it's only used in the `WorkspaceCreate`
view, which passes by position as that's the behaviour of a `CreateView` and
the test code that passed `data` by `kwarg` was not robust enough to detect
that the data it tried to supply was just being ignored in those tests.

The expanded `test_workspacecreateform_success` with multiple repos and not
choosing the first one was failing when `data` was passed by position.  This is
because the `__init__` was partially ignoring this parameter, unintentionally.
As the default values were the same as the POSTed data when only one option was
provided, the test passed anyway.

Expand `test_workspacecreateform_success` to have two repos and select the
second provided, to validate that `self.data` is actually being used.

Briefly document the intent of each test.