Removed unused parameter to `WorkspaceCreate` form.
`WorkspaceCreate` was passed a `project` parameter but didn't do anything with it.
Removing it to tidy up.

The parent view needs `self.project` for when it creates a new Workspace object
in `form_valid()`, but the form doesn't need it. Perhaps at some point it was the
form code that saved the object and it did need the `project`.