Don't reset repo on invalid `WorkspaceCreateForm`.
`WorkspaceCreateForm` constructs repo and branch `ChoiceFields` dynamically
based on the state of GitHub at request-time. The code that does this was not
doing anything to maintain the initial value from any submitted form POST data.
The Django machinery does not do this automatically for these fields as they
are dynamically constructed. Therefore if an invalid form was submitted (e.g.,
with a name that is not a valid slug), these fields were reverting to their
default value of the first element in the list, which is bad UX as they need to
reselect the repo and branch on validation failure.

Set the `initial` property on each field to the value from the POST data, or
`None` if not present.

`initial` from the `ChoiceField` API is not automatically respected by our
`form_select` components, so set the initial selected value in the template.

Make the branching logic clearer and add more comments.

Add a test case for the state of the unbound form (on first visiting), as
this was missing and none of the existing tests validated that the `ChoiceField`
`choices` are constructed correctly from `repos_with_branches`.