Add dummy invalid repo option.
This hints that the user needs to do something, and prevent auto-selecting the
first repo by default if the user does not change the field.

We expect that this can never be submiited because the branch choices will not
be populated on an unbound form. If it somehow were, the `clean()` already
validates that the repo URL can be found in `repos_with_branches`, which won't
be the case for our dummy value.