Improve unit tests of `WorkspaceCreateForm`.
Add a test case for the state of the unbound form (on first visiting), as none
of the existing tests validated that the `ChoiceField` `choices` are
constructed correctly from `repos_with_branches`.

Expand `test_workspacecreateform_success` to, like the new unbound test, have
two repos each with multiple branches, and don't select the first ones, to
validate that `self.data` is being used to construct the field options and
`cleaned_data`. This reveals a bug: if `data` is passed by position our
`__init__` partially ignores it as it refers to `kwargs` not `self.data` as it
should. Adjust the test to pass data by kwarg for now, a later commit in the
same PR will resolve this.

Briefly document the intent of each test in both name and docstring.