Refer to `form.data`, not `kwargs`.
POSTed data may be provided to a `Form` `__init__` function as either the first
positional argument or as the `data` key-word argument . Therefore we should
access it through `self.data`, not through `kwargs`. Otherwise `data` will be
partially ignored if passed by position, which doesn't follow the `Form` API
and is surprising.

This hasn't been noticed so far because it's only used in the `WorkspaceCreate`
view, which passes by position as that's the behaviour of a `CreateView` and
the test code that passed `data` by `kwarg` was not robust enough to detect
that the data it tried to supply was just being ignored in those tests.

A previous commit in this branch expanded `test_workspacecreateform_success`
with multiple repos and passed the second one in `data`. This failed when
passed by position, because the `__init__` was partially ignoring this
parameter, unintentionally.  Before this branch as the default values were the
same as the POSTed data when only one option was provided, the test passed
anyway. Therefore the previous commit in this branch had to change test code to
pass by position. This commit reverts that change. This provides some limited
assurance that the code respects the `Form` API, by including test cases with
both approaches to providing `data`.