Simplify `UserManager`, `get_or_create_user`, `UserQuerySet`.
Remove `UserManager.create_user` that invoked Django's `UserManager`'s private
function_create_user`. Retain the normalization and password hashing in
`.create()`. This is simpler and clearer.

Change `UserManager` to extend `Manager` instead of Django's `UserManager`. We
don't need the other UserManager public functions `create_user` and
`create_superuser` (which relate to the `is_staff` and `is_superuser` fields
that we removed in #4625), or `with_perm` as we don't use Django's `Permission`
model. This is simpler and clearer.

Remove the non-standard `get_or_create_user` and replace in client-code with
the standard `objects.get_or_create` which it replicated for `create_user`.
This is simpler and more standard and less code to maintain.

Replaced `UserQuerySet` with `User.Meta.ordering`, which is the standard Django
approach for defining default ordering. Updated most clients to use
`User.objects.all()` instead of the custom order_by_name from `UserQuerySet`.
Added a third sort criterion by 'username', as fullname is not guaranteed to be
unique.