Remove `User.is_active`.
Nothing referred to this field except one create function. It most likely
exists to mirror Django's `AbstractUser`, but if we no longer use it we should
remove it.

Note that the references to `is_user_active` in some templates are referring to
a variable that they define which is not related to this property of the `User`
model.