Don't patch `form.cleaned_data`
We patched `form.cleaned_data` because we wanted to reach the form's
`clean` method. Form and field validation failed before this method was
called, because `form.branch` is a `ChoiceField`. We no longer need to
reach the form's `clean` method, so we no longer need to patch
`form.cleaned_data`.