Remove references to IG processes.
The site can't know what the current processes are. So don't describe them.

Mentioning and linking to the policy is useful as it provides breadcrumbs to
the primary contextual motivation for the dashboard existing, without strong
assumptions about process at any given point in time that the site/page can't
really know.