Verify that fakes have the right methods.
We offer fake versions of key APIs for testing to avoid the expense of using
the real API and to reduce reliance on mocks. Ensuring that the methods and
their signatures match between the fake and real APIs helps keep them in sync.

Fix one signature that (without consequence) was slightly different in the
fake.