Parametrize test
We parametrize the test to demonstrate that the filter operates on
global *and* local roles. Ideally, the test wouldn't depend on us
knowing that `OutputPublisher` is a global role and `ProjectDeveloper`
is a local role, but making it so would add (arguably unnecessary)
complexity. To clarify our intent, we assign descriptive IDs to each
parameter set using `pytest.param`.

Co-authored-by: Mike Kelly <mike.kelly@thedatalab.org>