Group test cases by how they use the remote API.
This makes it clearer how the tests behave and makes it much easier to run
locally the tests that don't need permissions on a private remote repo. This
can be useful if you don't have that configured, for example.

Apply `enable_network` class-wise rather than function-wise as this makes it
clearer what's distinct about each test from their signature. Note the `github_api`
parameter is accessed in the tests so has to be applied function-wise.