Update backend state on GET requests
As described in the linked ticket, previously we only updated this
state on POST requests. This meant that when we entered database
maintenance mode this would only become visible to users once a job was
actually running. This has repeatedly caused minor confusion and
annoyance.

We were already updating the "last seen" stats for the backend on every
GET, so this just extends this slightly.

Closes #4605