Increase bot crawl delay
Tell bots to hit our site no more than one every two seconds.

We're currently seeing higher than normal traffic on the website and we think
that may be due to a high number of bots. This should reduce the impact of any
bot traffic and reduce the load on our server.