Get or create release
Check all the data passed to the create_release endpoint and return
an existing release, if one matches. Allow for the endpoint to be
called by a different user (in the event that a release is created by
one output-checker, and then a re-release attempt is made by another
output-checker). This can happen if not all files successfully upload
from Airlock.

The create_release is wrapped in transaction.atomic, so if we found
and existing matching release, we should be able to safely assume that
all the request files have already been created.