Check existing releases by ID
Releases should always be created by Airlock now, with an airlock
id which is used as the release ID. We can filter for existing
releases on release ID, and avoid filtering with requested_files,
which is a json field and filtering works differently.

If we find an existing release with the airlock ID, we check the
shas of the requested_files match.