Use `urllib3` v2
Fixes #4715.

We reverted to urllib v1 in #4806 to try and stop errors originating
from OpenTelemetry's use of `urllib3`, via `requests`.

Apparently these errors did not occur with `urllib3` v1 previously, but
that is no longer the case, and we'll have to find another solution to
stop the errors, as in #4839.