Send notifications for Airlock approved events
Airlock now releases files asynchronously, so an Approved request
means that the release request has been approved and files will be
uploaded in due course. Only when the last file is uploaded will
the request change to Released. It therefore makes sense to send
users notifications of the change of status to Approved.

This sends an email to the request author, adds a comment to the
GitHub issue, and posts a notification to slack about the status
change.