Update slack for returned requests
Previously we didn't notifiy slack for returned requests, on the
basis that the author gets emailed anyway, and the output-checker
already know that they returned it. However, it's useful to see
the progress of a request if all notifications come to slack, either
via issue creation/close or via an explicit post to slack. Currently
the returned request is the only status change that doesn't update
slack, so other output checkers would need to go to the issue or to
L4 in order to check if someone else has returned it. This is more
so now that a request can be returned before it's been fully
approved.