Add `SiteAlert` model for site-wide alert messages.
The intention is to display these to authenticated users on each page so that
they can be notified of ongoing site-wide issues like extended downtime.

One could imagine possible future extensions:
- `is_active` flag to draft alerts without making them visible and hide them
  without deleting them.
- `priority` field to override the ordering.
- `valid_from`/`valid_until` fields to schedule alert visibility.
- Fields to control visibility among users and/or restrict to certain pages
- Dismissible alerts persisting some state in session.

`paired_field` tests disabled because it's valid for a user to create or update
a site alert then for that user to be deleted, and it's still useful to know
_when_ the alert was touched, so I've used `SET_NULL` on the _by fields and
don't have any constraint relating the fields.