Remove template warning for no staff search query
Fixes #4897.

(This is only a single warning fix, but we agreed there that this was
all that was necessary right now to close the issue.)

When no search query was entered, `value` was not set at all, but
specified for use in the template.

Fix this by setting to an empty default.