Remove use of `pytest-subtests`
Fixes #4834.

Refactor the single collection of tests that used `pytest-subtests` and
use `parametrize()` instead.

If we used the package more, maybe we could justify leaving it be, but
it seems a good candidate for dependency removal.

It also unlocks removing `attrs` (#4860).

This also reworks the tests to also ensure all the icons are actually
tested, as spotted by Mike during review.

Co-authored-by: Mike Kelly <mike.kelly@thedatalab.org>