Remove use of `pytest-subtests`
Fixes #4834.

Refactor the single collection of tests that used `pytest-subtests` and
use `parametrize()` instead.

If we used the package more, maybe we could justify leaving it be, but
it seems a good candidate for dependency removal.

It also unlocks removing `attrs` (#4860).

This does seem to slow the test down very slightly (running this set of
tests alone takes about 6 seconds compared to 5 before this commit,
though most of this is the pytest/xdist startup time).