Remove Interactive login form
We're removing OpenSAFELY Interactive and this is part of that
functionality.

This prevents existing Interactive users from logging in.

It removes:

* the `login-with-url` URL pattern
* the associated `LoginWithURL` view
* the form handling on the `Login` view that took a user's email and
  sent them a login email
* the code to send that login email
* the section of the login template that refers to
* the tests that check the functionality of logging in by the emailed
  link

This restores the code roughly as it was prior to adding the Interactive
login-via-URL process in 2b09051cac5614d0b0d2bf3a99b72c08fdd08824

We retain the fix introduced in 130691df4c0a4e51ec20bcc892aa92345bdf2f17 (#2766).

One difference from the pre-Interactive code is that login now uses a
class-based view instead of using a view function, which helps minimise
the changes (for example, there are several tests that would need
changing), and keeps that view consistent with the others in the same
module that are all class-based.

***

Slightly unrelated, but while cleaning up, this also removes the
`TokenLoginForm()` from `Login`'s context. This was originally added in
6904755dbc765de3671cb1bd5f39791d2af98d3a.

This piece of context data is no longer needed given that Airlock
provides access to view output now, and is not even used in the
template. The use of this `TokenLoginForm()` was removed from the
template in 48f9565645700f6479ce108524cad1f73b826672, but this context
was overlooked. See also #4939 which proposes removing of
`LoginWithToken`.