Try all airlock notification actions in turn
Try each notify action in turn, catch any exceptions, and return
them as a list so that we can report them. This means that an
unexpected exception in one notification action won't prevent others
being tried, e.g. if we get a Github error updating an issue, we'll still
send a notification to slack.