Display hint strip in development environment.
If running in development, display a soothing green strip at the top of each
page to reassure dev users that they are not accidentally on the real live
production site.

Don't display the pre-existing red one on staff pages unless in production.
Staff users don't need to be warned about the risks of changing production data
with the strip when unless they're actually in the production environment. Also
re-name the file to make its purpose a little clearer.