Remove InteractiveReporter role
In order to support a migration to remove previous assignments of this
role to users, this role class must remain (albeit with no permissions).

Once this future migration has been applied and squashed out of
existence, the references to this class will be removed and so this
vestigial role class can be removed.