Increase sentry cron failure threshold
We get sporadic single failures of the dump_db job.

This generates a lot of noise of new issues in Sentry.

They tend to be singular failures (i.e. followed by a successful run),
by making the failure threshold 2 we will only get notified if two in a row fail.