Update to Django 5.1.7
Dependabot is unable to create this update, but `pip-compile` can.

In Dependabot's case, it's complaining about `psycopg[binary]`. We could
update that in the same commit/pull request, but it's perhaps good to be
cautious, because there's an open issue for `psycopg` that claims
performance issues with the latest minor version:

psycopg/psycopg#888

That issue may or may not be applicable; it possibly could be a
misreporting (although a few people have commented), or dependent on the
actual circumstances in which `psycopg` is used.

In any case, it's probably prudent to update only Django, seeing as we
can.