Remove user.is_interactive_only
is_interactive_only is strongly coupled to the Interactive Reporter role
and so does not function properly without it.

We cannot assert is_interactive_only as always False for there are users
who were (by the previous definition) "interactive only" and so we should
not show them the bits of the templates that we previously asserted
should not be shown to "interactive only" users.

This commit is added after the migration to remove the role from users
such that it is deployed together and the previously interactive-only
but now role-less users only see the things they are supposed to.