Remove `test_releaseapi_post_success_for_html`.
This test has a docstring referring to a conditional relating to
`AnalysisRequest` and whether the file is HTML or not. If we refer back to
commit d9154655ce4d809e306f44d0b1886d96de1990c1 we can see a conditional
involving `if ext == ".html"` and `AnalysisRequest`. That no longer exists in
`jobserver/api/releases.py::ReleaseAPI.post`. So I don't think we need this
test anymore. It is identical to `test_releaseapi_post_success` apart from the
filename.