Squash `redirects` migrations; remove `interactive` dependency.
We've removed `interactive.models.AnalysisRequest` in #4962. Migration
0001_initial introduced things that are then reverted in 0003. That's not
interesting, and the dependency on `interactive` was blocking us removing the
`interactive` migrations and app.

Squashing allows us to easily produce a set of valid migrations for the
`redirects` app to replace the existing ones that don't include the dependency.

I've re-named 0001 with git and changed to the content of the generate squashed
migration to make the git history slightly easier to follow.