Re-name tests.
The difference between these tests is that one of them passes the optional
`copilot` and `application_url` parameters to the command and one of them
doesn't. Change the names to reflect that.

I don't see how these differences should lead to a project being called
"interactive" or "standard". And in any case, we're sunsetting Interactive
(docs/adr/0029-sunset-osi-interactive.md), but it's still valid for a `Project`
to have these attributes, or not; and the command still has those options. We
will still want this for testing projects, for example.