Backfill fullnames
Since f88e84280d78e6abbca166b18232116d9398509d (Apr 2023) the user sign
up form has required a `fullname` but at the time this wasn't enforced
in the model.

This left 27 (at the time of writing) vestigial nameless users.

In advance of enforcing the non-nullness of User.fullname in the model
this commit backfills a full name for these users.

First priority is known system accounts for which we declare a name.
Next is looking for applications from a given user which have a full
name populated and taking that.
Next is to try and parse firstname.lastname@[domain] email addresses.
Finally if these fail then populate with the username.