Mark long integration test as `slow_test`.
This tests takes around 0.6s to run locally for me, about three times as long
as any other test. Marking it marginally speeds up `just test`, shortening some
feedback loops when doing local development. It's a very long integration test
with many steps. It's testing the applications app that we're not currently
developing and is planned for removal, so it's not likely to harm developers to
move this to CI testing (or `just test-ci`).