Simplify out `RemoveField` from squash migrations.
Remove the `RemoveField` operation and the field added in the `CreateModel`
operation that it relates to. This has the same end result but is in fewer
operations.  Maybe it make our tests marginally faster.

Django won't make these kind of simplifications automatically, too hard to
reason about. But we can, if they're not too much effort.