Comment on dependency purpose and usage.
We're oftentimes wondering "what is this package for? roughly where is it
used?". Here are some comments explaining the current state of play.

They could possibly get slightly out of date if someone uses a dependency in a
new place. That's probably not too damaging and won't happen too often and can
be weighed against the usefulness of having at least some information readily
available.

These comments are also useful when checking for childless dependencies. I
needed this information to do that process, so I'm writing down these notes for
future reference.