Replace use of `incuna_mail` with wrapper function.
The only thing we use the `incuna_mail` dependency for is wrapping creating and
sending a `EmailMultiAlternatives`. This is just a few lines of code we can
write ourselves, so let's do that instead. The version added here is simpler
than the `incuna_mail` wrapper and handles all our present use cases.

This will allow us to simplify by removing an unnecessary dependency. There are
also some marginal benefits to reducing our attack surface if a package update
were compromised and risks of future incompatibility.