Use update-dependencies-action for dependencies
Instead of Dependabot, which has been inconsistent in its behaviour in
various ways.

See, for example, #4412.

This was previously checking daily for `pipeline` updates, but in
Airlock, these are only checked weekly along with the dependencies
themselves. For simplicity, let's do the same (otherwise we need some
kind of condition checking the weekday, and having different behaviour
for scheduled/non-scheduled workflow runs, and scheduled runs depending
on the weekday).

If a `pipeline` update is needed sooner, the workflow can always be run
manually.