Add user.has_any_roles property
This replaces the `all_roles` property in the two templates where it was used.

Note: in the job detail page we're effectively using this property to determine
whether someone can view job metrics by calculating if someone's account is
enabled. We should be careful about using the property in that way more widely.