Remove `Project.application_url` and migrate.
This was never used so we should remove it. No references to the field remain
so it should be safe to remove.

As part of the Sunsetting OSI initiative (see
`docs/adr/0029-sunset-osi-interactive.md`), we removed all OpenSAFELY
Interactive functionality from Job Server. The `application_url` field on a
`Project` was added specifically for Interactive project applications. However,
the standard application process was actually used, so this field was never
populated. We have no intention of using it in future.