Replace `pytest-network` with `pytest-socket`.
The two packages do broadly the same thing -- ensure tests don't access the
network. `pytest-network` is not being actively developed, `pytest-socket` is.
We probably benefit only marginally from updates to this package, but we may as
well.

We could have instead written our own code to disable socket access to remove
the dependency, but it's just complicated enough that it seems worth a
development dependency (which has a lower cost-benefit barrier than a
production dependency).