Ignore labels for airlock issues if they don't exist for the repo
When a new release request is submitted on Airlock, we create a new
issue for it. By default, this has either the "internal" or "external"
label applied. However, non-opensafely output-checking repos may not
have these labels, and if we try to create an issue with a non-existent
label, we get a 403 from the github API.  We now check for labels on
the relevant repo, and ignore any that don't already exist.