Use timestamps from RAP Controller in backend stats
We track "last contact" statistics for each backend which we use to
confirm that the backend is available and update status pages and issues
alerts appropriately. But since the RAP Controller now sits between the
backend and Job Server, contact from the controller is no longer
evidence of the health of the backend.

To address this the RAP Controller now tracks last contact times for the
agent in each backend and passes these up to Job Server. This PR makes
Job Server use those timestamps, rather than the current time, as a
record of when each backend was known to be alive. The rest of the
status checking mechanisms can remain unchanged.

See:
 * https://github.com/opensafely-core/job-runner/issues/1037
 * https://github.com/opensafely-core/job-runner/pull/1049