Improve performance of JobRequests endpoint
This endpoint is intended to return all "active" JobRequests, which
means all JobRequests with active jobs and any JobRequests with no jobs
at all (which means they have been recently created and not yet
acknowledged by the RAP Controller).

Previously it did this using a JOIN between the JobRequest and Job
tables and the a DISTINCT to squash the results back down to one row per
JobRequest. However we can also do the same thing using subqueries and
this turns out to be significantly faster. Running the equivalent
queries through the management console shows a speed-up of somewhere
between 25x and 30x.