Only call `refresh_from_db()` when actually needed
This endpoint is having some performance issues when called with very
large numbers of jobs and so reducing unnecessary work seems like a good
idea. There's not an easy way to determine the production impact here
ahead of time but:

 * this is obviously safe because if we don't take the `newly_completed`
   branch then the `job` object is immediately discarded;

 * it can't _hurt_ performance to not call `refresh_from_db()`.

So overall this seems worth trying.

Also worth noting that this call was only [introduced][1] to deal with
problems with the notifications, so it's only ever been needed for this
branch.

Slack discussion at:
https://bennettoxford.slack.com/archives/C069SADHP1Q/p1750839710327239?thread_ts=1750768496.173799&cid=C069SADHP1Q

[1]: https://github.com/opensafely-core/job-server/pull/345